+++ admin-part/src/main/kotlin/Build.kt
@@ -19,7 +19,7 @@ import com.epam.drill.plugins.test2code.api.*
 import com.epam.drill.plugins.test2code.common.api.*
 import com.epam.drill.plugins.test2code.coverage.*
 import com.epam.drill.plugins.test2code.group.*
-import com.epam.kodux.*
+import com.epam.dsm.*
 import kotlinx.collections.immutable.*
 import kotlinx.serialization.*
