+++ admin-part/src/main/kotlin/Plugin.kt
@@ -26,7 +26,7 @@ import com.epam.drill.plugins.test2code.coverage.*
 import com.epam.drill.plugins.test2code.group.*
 import com.epam.drill.plugins.test2code.storage.*
 import com.epam.drill.plugins.test2code.util.*
-import com.epam.kodux.*
+import com.epam.dsm.*
 import com.epam.kodux.util.*
 import com.github.luben.zstd.*
 import kotlinx.atomicfu.*
