+++ gradle.properties
@@ -19,10 +19,12 @@ xodusVersion=1.3.232
 licenseVersion=0.15.0
 benchmarkVersion=0.3.0
 kotlinNoarg=1.4.21
-zstdVersion=1.4.8-1
+zstdVersion=1.5.0-4
+testContainerVersion=1.16.2
 
 bcelVersion=6.5.0
 jacocoVersion=0.8.7
 
 drillApiVersion=0.8.0-4
-drillAdminVersion=0.8.0-+
+drillAdminVersion=0.8.0.38.postgresql.2
+drillDsmVersion=0.1.3
